Update .gitignore even if srcdir != builddir
When srcdir != builddir, autotools sets VPATH = @srcdir@.
due to this and unfortunately the top directory has same name to
liblangtag, where contains the library's source code, make confuses
the top directory is the target directory for 'liblangtag'.
.gitignore is already available since building is performed at the top
therefore .gitignore considered available.

To avoid this confusion on make, abs_srcdir has to be used for target.